Add `clone` to DocumentPartRoot
You can now clone a DocumentPartRoot, and it will a) clone the
underlying document or document_fragment, b) clone all contained `Part`
references from the original to the clone, and c) return you the new
DocumentPartRoot. To allow access to the underlying cloned Node tree, I
also added a DocumentPartRoot.root, which returns the owning Document
or DocumentFragment.

So:
  const root = document.getPartRoot();
  { ... add NodePart's and ChildNodePart's to root ... }
  const clonedPartRoot = root.clone();
  const clonedDocument = clonedPartRoot.root;

To make this work, during cloning (via partRoot.clone), references
are kept in NodeCloningData that point from cloned Nodes to their
clones, and from cloned PartRoots to their clones. Also, a "queue"
of to-be-cloned Parts, keyed by the PartRoot they're waiting for,
is used to handle the common case that the PartRoot for one part
might not be cloned before a Part that refers to it. An example is:

<div id=parent>
  <div id=previous_sibling></div>
  <div id=middle></div>
  <div id=next_sibling></div>
</div>

Assume there is a ChildNodePart with previous_sibling/next_sibling,
and a NodePart pointing to middle which has ChildNodePart as its
PartRoot. During the clone of this tree, the ChildNodePart can only
be cloned once parent, previous_sibling, and next_sibling have all
been cloned. However, middle (and its NodePart) will get reached by
the clone operation before next_sibling. So NodePart here needs to
be queued and retried once ChildNodePart is cloned.

Bug: 1453291
Change-Id: I4a85c3ee02547627bcf73ed610310c742238f12c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4671112
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168260}
NOKEYCHECK=True
GitOrigin-RevId: 5d60017dba57e86d477634812e1340127734f8a7