Change WebLocalFrameImpl to use Member to hold the current history item.
Previously, the current history item was owned by the WebLocalFrameImpl
as a WebHistoryItem, which is intended to an opaque handle to a
HistoryItem that non-Oilpan aware code can use. Internally, it uses a
Persistent to ensure that the garbage collected HistoryItem remains
live.

However, this means that a garbage collection pass can't clean these
objects up in a single pass once they are no longer unreachable: the
Persistent artificially marks the rest of the object graph as reachable
until the WebLocalFrameImpl is finalized, and only then will subsequent
garbage collection passes reclaim the rest of the objects.

Change WebLocalFrameImpl to use Member<HistoryItem> to keep the history
item live. Since WebHistoryItem had some functionality on top of
HistoryItem, e.g.  specifying the target frame name and PageState
serialization, this functionality has been folded into HistoryItem
itself. WebHistoryItem is now just a thin wrapper around HistoryItem for
the Blink public API.

This also allows some additional simplifications, i.e. the code can just
work with WTF::Vector and std::vector directly rather than going through
WebVector. There may be other simplifications that are possible in the
future, e.g. not going through WebHTTPBody.

One quirk that this CL preserves is that GetReferencedFilePaths() has
two implementations with slightly different behavior: one on HistoryItem
and one on WebHistoryItem. Ever since WebHistoryItem was implemented in
https://crev.com/db4a9b4108635b3678c3f9fd5bdd1f98001db216, it returned
more paths than the HistoryItem-equivalent of GetReferencedFilePaths().
From that CL's description, it looks like it was to make sure that the
browser granted file access to any referenced files in a restored tab.
While it is unclear if this is actually still necessary, it is safer to
preserve the existing behavior for now.

Bug: 1458809
Change-Id: I1225dca87655a052fe349f48691e1843ca391010
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4664558
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168295}
NOKEYCHECK=True
GitOrigin-RevId: f5aebf49678bc27eb1b1a61bc30775bacc325fa0