Fenced frames: Navigate to about:blank when config attr is set to null.
Note: This does not affect the initial state of the config member, and
only takes affect when it is explicitly set to null. Continuation of
https://crrev.com/c/4341914.

click `Add FencedFrame`, and enter
`document.querySelector('fencedframe').config = null` in the console;
the frame should display a blank page.

Change-Id: I07dbcefaea25bea74f4a3ad6880800036224338a
Test: 1. Added a browser test 2. Navigate to https://mparch.glitch.me,
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4636931
Commit-Queue: Andrew Verge <averge@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168356}
NOKEYCHECK=True
GitOrigin-RevId: 9fd5755e176f377e2b4b96f7dfcc469d2e7d93a1