Implement state transition for AudioContext.setSinkId
This CL updates the implementation of AudioContext to follow the
spec algorithm of the state transition from setSinkid() method.

https://webaudio.github.io/web-audio-api/#dom-audiocontext-setsinkid-domstring-or-audiosinkoptions-sinkid

The relevant web test is also updated and the failing expectation file
is removed. The actual audio output on MacOS is tested by:
https://googlechromelabs.github.io/web-audio-samples/tests/setsinkid/

Bug: 1439947
Change-Id: I56d8c89272d66b08502203a846c65231a72e8549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4670272
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168371}
NOKEYCHECK=True
GitOrigin-RevId: ffc8d65e4db54a8c7d19f4f2350fe360ede7fe27