"Correct" the class hierarchy for DOM Parts
The prior hierarchy was a bit wonky:

            PartRoot (getParts, clone)
             /    \
DocumentPartRoot  Part (root, metadata)
                 /    \
     NodePart (node)   ChildNodePart (prev/next/etc)

This was funny because a Part was a PartRoot, but kind of not always,
since a NodePart isn't really a PartRoot. That hierarchy was used
to avoid the inconvenience that multiple inheritance isn't a thing
for JS. This CL updates the class hierarchy to implement PartRoot as
a JS interface mixin, with the C++ implementation in a single PartRoot
base class shared by DocumentPartRoot and ChildNodePart. So the new
C++ class hierarchy looks like this:

          Part         PartRoot
         /    \        /      \
  NodePart   ChildNodePart   DocumentPartRoot

This should make future additions (such as the ability to clone
ChildNodeParts) less painful.

Bug: 1453291
Change-Id: I1c3e820fc6d76aa56dcf3a2e0f26cfc6ab30f3e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673702
Reviewed-by: David Baron <dbaron@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168448}
NOKEYCHECK=True
GitOrigin-RevId: fab05ecf6035119199a9e17a94c274129d0f6cad