Add Zstandard Content-Encoding support
This CL pulls the zstd library[1] into third_party/zstd via DEPS, and
adds Accept-Encoding and Content-Encoding support.

This means that Chrome can now send "zstd" as a token in the
Accept-Encoding request header, and decompress with Zstandard[2]
when the "zstd" token is seen in the Content-Encoding response
header. This will be enabled only when |enable_zstd| is set to true
in the NetworkContextParams.

[1] https://github.com/facebook/zstd
[2] https://datatracker.ietf.org/doc/html/rfc8878

Binary-Size: Added the zstd library.
Fuchsia-Binary-Size: Added the zstd library.
Bug: 1246971
Change-Id: I793335ea927d2a7332f919fbc443a235d9708939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4617470
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168488}
NOKEYCHECK=True
GitOrigin-RevId: bafe20fdae189e50d59ae6a9628712644a92e66a