Count UseCounter regardless of the feature flag
The previous CL [1] introduced a UseCounter for "port overflow", but
it counts only when the feature flag is enabled. The flag is disabled
by default. The UseCounter currently remains at zero [2].

This CL fixes this behavior so it counts regardless of the feature
flag.

[1] https://crrev.com/c/4252309
[2] https://chromestatus.com/metrics/feature/timeline/popularity/4503

Bug: 1416017
Change-Id: I00662cfbd663bbd584fc7d6392bf08bf91d02ab3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4667092
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168568}
NOKEYCHECK=True
GitOrigin-RevId: b09d6224d569895ce51269c1a4d84830e43058e1