Store the COOP restrict crossOriginIsolated bit in policy container.
In https://crrev.com/c/4479858, we introduced a new boolean that
constrains whether an execution context can be considered
crossOriginIsolated. It is computed by the browser, depending on whether
the COOP origin matches the actual origin. This boolean is computed and
sent to the renderer at window creation time, or at commit time for
subframes. It causes breakage such as b/288962508 (Google
only) because SharedArrayBuffer's constructor sometimes needs to be
enabled earlier than CommitNavigation, which causes SharedArrayBuffer to
be undefined even though crossOriginIsolated is true.

This CL instead puts the boolean inside the policy container and lets it
handles inheritance.

Bug: 1385827
Change-Id: I1330c30ad056d9f94ee7b7747eac68ae47b56acc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4666003
Commit-Queue: Jonathan Hao <phao@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168653}
NOKEYCHECK=True
GitOrigin-RevId: 81558f63013359a271b0fb67df9c724b075d840e