Rename Attribution registration types
This CL was created to be able to use the`OsRegistrationType`
in a non-OS setting in: https://chromium-review.googlesource.com/c/chromium/src/+/4639245

To do so, we first rename `RegistrationType` to the more accurate
`RegistrationEligibility`.

Which enables us to then rename `OsRegistrationType` to
`RegistrationType`.

Bug: 1374121
Change-Id: Ib8c4a0851422e56a9525e7ee8ae3472f5a841753
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4650271
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168676}
NOKEYCHECK=True
GitOrigin-RevId: 08dcf2bb417c49b132ede5fb2de2834991cf531d