gfx::HDRMetadata: Remove gfx::HDRMode
The gfx::HDRMode enum was added to signal that a surface would use
values outside of the [0, 1] range. This was the wrong way to solve
the problem. The better solution is to use extended_range in the
existing gfx::HDRMetadata.

Remove the extensive plumbing of gfx::HDRMode.

While removing that plumbing, change the usually-adjacent
absl::<gfx::HDRMetadata> to just be a gfx::HDRMetadata. This is because
we have changed each of the members of gfx::HDRMetadata to be optional
(instead of having the whole structure be optional).

Bug: 1446302
Change-Id: I8c6c77acf68b86084910e023cc4e06e216a15175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614745
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168738}
NOKEYCHECK=True
GitOrigin-RevId: ffb089a0f9dbd6a0137170feaea53046fffd6c02