Make source of performance entry weakly reference DomWindow object
Currently the source of a performance entry keeps a member pointer to
to the corresponding DomWindow object which keeps it alive. This causes
memory usage issue.

This CL is to make the source weakly reference the DomWindow object so
it can be GC'ed.

Bug: 1458831
Change-Id: Ia3c5cb00302b76af41cc3d024a146df45647f2f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4666321
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168820}
NOKEYCHECK=True
GitOrigin-RevId: 7096891340b9fc8619e4e1614edc36899c0fc042