Have automatic beacons sent from ctrl+click.
When a user ctrl+clicks a link to open a page in a new tab, the signals
that are currently checked for automatic beacons (initiator user
activation + initiator activation status in the navigation request
params) to determine if there is user activation are not set. This
causes a beacon to fail to send.

This CL modifies the automatic beacons user activation check to also
check if the `was_activated` flag was set in the commit params. This
flag is set explicitly for cases like ctrl+clicking a link[1].

This CL also adds a new internal WPT to test this case. We are using an
internal WPT instead of an external WPT to get access to the
`eventSender` class, which supports sending ctrl/cmd+click events.
FLEDGE-related helpers from the fenced frame external WPT folder are
brought in so that we can run ad auctions and check automatic beacon
reporting.

[1]https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/navigation_request.cc;l=3981-3986;drc=892166276bba156857c6321a1c5035c83200b55b

Bug: 1463784
Change-Id: I82121144af0365466043f0b96cab6d8e530accb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4670828
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168821}
NOKEYCHECK=True
GitOrigin-RevId: 3b7cb9a360242c6628ec19faf24ad68c805d73af