[A11y] Support rowindex attribute on rows
The aria-rowindex attribute can be specified on the row or the first
cell in a row.

Also cleans up some unnecessary caching of row and column index on
the Blink side. It's no longer necessary because table info is now
computed in Blink.

Bug: 1191936
Change-Id: Ibc8e617c05db1dc6f240bd574055b4233258f547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4661211
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Jocelyn Tran <jocelyntran@google.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1168920}
NOKEYCHECK=True
GitOrigin-RevId: 87676c7a2af3fcf69d3b92ea39d743372f44a255