SmartCard: add event count to GetStatusChange
This is an undocumented API but both Windows and PCSClite
implement it. PCSClite does document it, luckily. It's also not part of
the PC/SC spec.

Moretheless, one does need to provide this information on Windows
in order for that function to work correctly, otherwise it will always
return immediately as the number of expected events (zero) won't match
the actual event count for that reader.

This information is "smuggled" in the upper two bytes of the
dwEventState and dwCurrentState member variables of the
SCARD_READERSTATE struct.

Bug: 1458318
Change-Id: Ibb6f2e3d176af8c779cce9d59d2c30a542d0735d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4672938
Commit-Queue: Daniel d'Andrada <dandrader@google.com>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Viktoriia Kovalova <vkovalova@google.com>
Cr-Commit-Position: refs/heads/main@{#1169286}
NOKEYCHECK=True
GitOrigin-RevId: 57813ccd1c2aa0715b4b90401598d3a99102ae94