Emit networkIdle following aborted client-side navigation
Currently, IdlenessDetector is kicked off at the time
DOMContentLoaded is dispatched. If the document parser
is aborted by client-side navigation, DOMContentLoaded
is never dispatched and hence IdlenessDetector is not
functional either. In case of a successful client-side
navigation, this may be the desired behavior, as all
events would then be dispatched by the succeeding page.
However, if the navigation is aborted, we're left in a
state where a CDP client that may rely on networkIdle
events is left in a suspended state. Since, technically,
an aborted navigation leads to networkIdle condition,
I think we should properly handle this case on the
idleness detector side.

Bug: 1462258
Change-Id: Idd6d44fde0757af22051f0ee6f3d6def64e10034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4679496
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1169616}
NOKEYCHECK=True
GitOrigin-RevId: 2d584aba6c742f3d4e7b1e7f8ea29ec334ec4477