dlp: Pass block/warn file name in ProgressStatus
Stores the name of the first blocked or warning file and passes it to
the Files app when the task completes with error / is paused. This is
needed in order to properly show the panel items (notifications) when a
single file is restricted by the policy - the notification text should
use that file's name to make it clear for the user which files to check.
This is only relevant for the single restricted file case, as with more
than one, the notification has a Review button which opens a dialog with
all the affected files listed. Therefore it's fine to always pass the
first warn/block file - which is the correct one for single file case,
and is ignored for multi file case, and is less complicated then
ensuring that this value is passed IFF a single file is restricted.

R=chromeos-dlp@google.com, lucmult@chromium.org

Fixed: b:290769150
Change-Id: Ia7534307d319806754dff927750121a4d51b08b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4677763
Reviewed-by: Luca Accorsi <accorsi@google.com>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Aya Elsayed <ayaelattar@chromium.org>
Auto-Submit: Aida Zolic <aidazolic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1169822}
NOKEYCHECK=True
GitOrigin-RevId: a78b77c6c56b197f9e92e3ec202a6349345e34dd