wei(wpt): Add cross origin tests and split into two files
There is a lot of useful behavior we'd like to test that
is running in the browser process so the only way
we can ultimately test that is by splitting our
WPTs into Android specific and platform agnostic tests.

The platform agnostic tests are basically testing
everything in the renderer before we make a mojo
call to the browser process.

I have kept the mock attester in the android specific
test for the moment because we don't get a response
back from the Android OS yet and it was convenient
to get "some" kind of response back.

Also added cross origin testing in this by opening
a separate window and requesting tokens from that.

It's possibly that we could end up with the
android-specific tests eventually just being the tests
for when an attester is available but I'd rather keep
these separate until we know more the spec'd token
details.

Bug: 1454404
Change-Id: I2019ed0f598aa51dff5cbb7bc7ed8e573014fae4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4669525
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1169983}
NOKEYCHECK=True
GitOrigin-RevId: b662cf5ae59f9283d96423d4d3cd999e9132d980