Only special-case document element outermost <svg> embedded via a frame
In LayoutSVGRoot::IsEmbeddedThroughFrameContainingSVGDocument(), it
wasn't considered whether the <svg> was the document element or not, and
thus the special layout rules would be applied for all outermost <svg>
elements. This lead to incorrect layout results being calculated in some
of these cases.

Return false from said function if the LayoutSVGRoot is not associated
with the document element.

Fixed: 1458399
Change-Id: I7be6eedb0d019fd59692764d9cf7357b3bc10427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4681730
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1170011}
NOKEYCHECK=True
GitOrigin-RevId: 60c02f6a60d604a21ccc521789bed36fb9c459e5