Fix overscroll delta calculation
InputHandler::ScrollUpdate was only counting overscroll deltas where
content did not move at all. This omitted deltas in cases
where content moved but not all of the input delta was consumed.

This is the cause of the flakiness for some WPTs so this patch updates
TestExpectations and removes an obsolete expected.txt file.

Bug: 1430921,1313311
Change-Id: I32248abcbd1217b78b13f69799e6ced7a9d35e31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4683472
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1170110}
NOKEYCHECK=True
GitOrigin-RevId: de485c3a4bef131de82ca761d0870807b94553fe