[Inspector] Skip adding source data for non-existent implicit nested groups
See the bug for the reproduction and issue description.

The crux of the issue lies in nested group rules[1] and how the parser notifies the observer when they encounter such rules.

Take the CSS below as example:
```
.foo {
    @supports(display: flex) {
        & {
           color: red;
        }
    }
}
```

While parsing the body of the `@supports` rule above, CSS parser notifies the observer that this might be the start of the implicit nested group with a zero length header range[2].

However, if there are no properties for this implicit nested group, CSS parser doesn't register it so such a group doesn't exist for the rule list got from the parser[3].

in this CL, to match the same behavior with the CSS parser while registering the source data, we check whether there are any properties parsed for this possible "implicit nested group".

If not, we skip adding  the rule itself and only take its child rules' data.

[1]: https://drafts.csswg.org/css-nesting-1/#nested-group-rules
[2]: https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:third_party/blink/renderer/core/css/parser/css_parser_impl.cc;l=2122;drc=255b4e7036f1326f2219bd547d3d6dcf76064870
[3]: https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:third_party/blink/renderer/core/css/parser/css_parser_impl.cc;l=2131;drc=255b4e7036f1326f2219bd547d3d6dcf76064870

Fixed: 1445720
Change-Id: Ifbd2e85873374929d0ceadf4fb5905eaf4ddd62b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4681759
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1170208}
NOKEYCHECK=True
GitOrigin-RevId: 31acf2db4c910a483631f29ed6556af02b93c6c5