Introduce did_use_shared_dictionary flags and UseCounters
This CL introduces did_use_shared_dictionary flags in the following
structures:
  - net::HttpResponseInfo
  - network::mojom::URLResponseHead
  - blink::WebURLResponse
  - blink::ResourceResponse

Also this CL introduces the following UseCounters:
  - SharedDictionaryUsed
  - SharedDictionaryUsedForNavigation
  - SharedDictionaryUsedForMainFrameNavigation
  - SharedDictionaryUsedForSubFrameNavigation
  - SharedDictionaryUsedForSubresource

To set this flag in HttpResponseInfo, this CL changes the
SharedDictionaryNetworkTransaction class to create a new
HttpResponseInfo with the flag set when it will use a shared dictionary.
And GetResponseInfo() will return the new HttpResponseInfo.

network::URLLoader::BuildResponseHead() sets the flag of
network::mojom::URLResponseHead. And blink::WebURLResponse::Create()
sets the flag of blink::ResourceResponse through blink::WebURLResponse.

When this flag was set, DocumentLoader::RecordUseCountersForCommit()
records the UseCounters for navigations, and
ResourceLoader::DidReceiveResponseInternal() records the UseCounters
for subresources.

Bug: 1413922
Change-Id: Icef7d1fb9db95783caae91ae0abb656e7bc3ed0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673767
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1170325}
NOKEYCHECK=True
GitOrigin-RevId: 65d1cdff2d6e7521789b52060e575e5d31f1b023