[execCommand] Support modifier+non-modifier in `KeyUp`/`keyDown`
This patch supports sending keys with modifiers in `keyUp` and
`keyDown` action sequences.

crrev.com/c/4654141 supported `keyUp` and `keyDown` for all
keys, but when they are combined, such as:
```
actions().keyDown(kAltKey)
         .keyDown('c')
         .keyUp('c')
         .keyUp('kAltKey)
         .send();
```
It's the keyboard driver's (`EventSender` in this case)
responsibility to keep track of the modifier key state, not
the `Widget`'s (which `EventSender` sends events to.)

This patch updates `EventSender` to keep track of the state.

This patch:
* Fixes 3 tests (6 sub tests.)
* Makes a WIP test in crrev.com/c/4679415 pass.
* Turns 1 timeout to a failure. The cause isn't clear but the
  test fails at wpt too:
https://wpt.fyi/results/uievents/interface?label=master&label=experimental&aligned&q=keyboard-accesskey-click-event

Bug: 893480, 1462196
Change-Id: If1cf6e0a290c9f060135b2a67322aed3dd559e5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4680842
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1170356}
NOKEYCHECK=True
GitOrigin-RevId: 79d2a31660fd22247ce3e94d0d52137669362bb0