Add some non-const getters to SVGTextPositioningElement to match IDL semantics.
This duplicates some getters on SVGTextPositioningElement to have
non-const semantics, to match the semantics of WebIDL, where the return
value is not const.  It retains the const versions because they are used
by C++ callers.

I believe this sort of code *should* be a compilation error, and I have
work in progress on a CL to the bindings code to make it one.  This is
one of two const-related changes in blink that needs to be made for that
CL.  (There are other type-related changes.)

Bug: 1464089
Change-Id: I2b2af70105f174cb741350e778de30c4b5215f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4687476
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Auto-Submit: David Baron <dbaron@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1170502}
NOKEYCHECK=True
GitOrigin-RevId: a1177e1ae9b8f987feb727d1b01a100c5941e1de