Add null check for PerformanceTiming before referencing.
Crash is reported at the place when PerformanceTiming is dereferenced.
See bug for details.

The PerformanceTiming could be nullptr if the Performance object
initiating it is a WorkerPerformance. In the crash report, the stack
trace shows the Performance::Mark task is posted to a worker.

Previous there was a null check for PerformanceTiming. It was removed
in CL https://chromium-review.googlesource.com/c/chromium/src/+/3656547 somehow.

Bug: 1458578
Change-Id: I4d44955fe207e3c4771da9a6be40cb59649787bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4672597
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1170644}
NOKEYCHECK=True
GitOrigin-RevId: 074c7a1eccf44435ca1c69f52321d0b912c5fd76