Fire the navigate event on the correct document on RenderDocument's LocalFrame<->LocalFrame swap
When RenderDocument is enabled, same-site cross-document navigations
will create a new LocalFrame instead of reusing the previous
LocalFrame. This means, when trying to run the navigate event on the
previous document during commit of the new document, we must reach
into the previous LocalFrame to get the correct document to fire the
event to, which this CL does.

Bug: 936696
Change-Id: I59b707e9bd932b10c0a1519b7c73bafe77fc522d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4685222
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171154}
NOKEYCHECK=True
GitOrigin-RevId: 47e45c140fb92f40fcb15456fd7dc0830ad7c2d9