Make recent SW running status visible in SW subresource loader.
This is a part of implementation to support
RouterRunningStatusCondition written in:
https://github.com/yoshisatoyanagisawa/service-worker-static-routing-api/blob/main/final-form.md

To allow the router evaluator to use the ServiceWorker running
status as a condition, this CL makes the
ServiceWorkerSubresourceLoader to have a recent ServiceWorker
running status.  To allow it to access the recent status,
there is an IPC to ask ServiceWorkerContainerHost to tell the
status.

Change-Id: I09941a8e9648f8097a285856f2df1d11e4732c5d
Bug: 1371756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4680729
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171597}
NOKEYCHECK=True
GitOrigin-RevId: 3ce36bacc6f07be18bc0c730a182d0031d7efa35