[mathml] Simplify NGMathPaddedLayoutAlgorithm::GatherChildren
Code coverage report indicates that OOF branch in GatherChildren is not
reached [1] even though it should be exercised by WPT test [2]. This is
because <mpadded> elements actually generates nodes that are
LayoutNGMathMLBlockWithAnonymousMrow, which in turn are either empty
or just contain an anonymous wrapper for <mpadded>'s children. Rename
the method GetContentAsAnonymousMrow to make this more explicit and
simplify it.

Bug: 1464154

[1] https://analysis.chromium.org/coverage/p/chromium/file?host=chromium.googlesource.com&project=chromium/src&ref=refs/heads/main&revision=c8a01d35ed430e49e8275f641f735e689a97c105&path=//third_party/blink/renderer/core/layout/ng/mathml/ng_math_padded_layout_algorithm.cc&platform=linux&test_suite_type=any&modifier_id=0#53
[2] mathml/relations/css-styling/out-of-flow/all-mathml-containers.html

Change-Id: I356acef2f15b7b04afd269de280a78b582d2a78d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4684183
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171625}
NOKEYCHECK=True
GitOrigin-RevId: 8806dade91f0f5dc9b58f0422ece41d23d72ddd7