dlp: Add FileBackedBlobFactory interface and browser side
This CL adds a new interface to register file backed blobs.
For frames, this interface is in sync with navigation and thus
can be used to retrieve the URL from which blobs are accessed.
This will be used to perform data leak prevention checks for
managed users having specific file related policies.

This CL only adds the interface and its browser-side implementation.
For the actual usage see children CLs in the relation chain.

See also comments and discussions in the original big CL https://crrev.com/c/4249117/159, from which this one was extracted.

More info:
- design doc (https://docs.google.com/document/d/1tiK7HcFivDIK288PlEtqxPWEhMy-A5FjPf0A8Gmx4g4/edit#bookmark=id.wbsjo0yog42m)
- slightly more detailed overview (http://docs/document/d/1FrYQYg7PqGmjbfsjjQx_AVcAW--KSyoYI-rL8fx-lu0?resourcekey=0-KJCn36cd-eFdWwIDhKuKuQ)

Bug: b/289974874
Test: out/Default/content_unittests --gtest_filter=FileBackedBlobFactoryImplTest.*
Change-Id: Iab35bbee1e85af4a935152d034220deb431c8139
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4664640
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Daniel Brinkers <brinky@google.com>
Commit-Queue: Luca Accorsi <accorsi@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171659}
NOKEYCHECK=True
GitOrigin-RevId: 41728a8dbbdbf640125decd2b57155a1170341c7