Speculative fix for perf regression from declarative DOM Parts
The only real change made in [1] is the set of feature flag
checks in HTMLConstructionSite::InsertComment() and
HTMLConstructionSite::ExecuteTask(). This CL moves that feature
flag check up to the constructor, and uses the presence of
pending_dom_parts_ as the feature check, which should hopefully
be faster than RuntimeEnabledFeatures.

In the process, I also removed InitFragmentParsing() and combined
it with the constructor call for HTMLConstructionSite. I'm hoping
that doesn't have any negative perf impact, and might even result
in a progression for fragment parsing?

Bug: 1453291
Change-Id: I2ae7f86e1d7b44c7f94b4a442e7b2cde821e0cb0
Fixed: 1465115
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4691342
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171762}
NOKEYCHECK=True
GitOrigin-RevId: 601f582c995f6df72dd801c90149abb376b1c322