Don't invalidate style if anchor href is reassigned the same value.
In these cases it is quite simple to avoid invalidating certain
pseudo-class states as well as the cached visited link cache.

In principle we could also skip some of this if the link is changed
in a way that has the same visited hash (e.g., by changing from
a relative to absolute URL with the same resolved URL), but that's
harder and unlikely to be a common case.

Fixed: 1464696
Change-Id: I6c18b60a3ea7b81f66adc9fa0b8e1f2ed2060d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4687934
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171836}
NOKEYCHECK=True
GitOrigin-RevId: ef51135d1e1f52f9c38b6ac816cc4f387af26231