[automation] Remove domQuerySelector
This API uses a method of communicating between processes that is not
supported with manifest v3. Since this method is part of a private API,
and is not used anywhere in the code base, it is safe to simply remove
it entirely.

Details on the automation API and its incremental deprecation can be
found here: go/chrome-automation-partial-deprecation

AX-Relnotes: n/a.
Bug: b:260589686
Test: Code search shows no usages, codebase builds and ChromeVox works.
Change-Id: Icfdf555d18616be822da618860bd4a344abcdd9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4679220
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Katie Dektar <katie@chromium.org>
Auto-Submit: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/main@{#1171914}
NOKEYCHECK=True
GitOrigin-RevId: 2f171fceccaa2b0f2b97d8fbd652d8635d272d62