[Viewport Segments API] Add Windows support for foldable PCs.
This patch enable Windows support for the Viewport Segments API on Windows
for foldables PCs (foldable OLED screen).
Due to lack of native API in Windows, the ecosystem gathered around the
following approach to inform the segments of the device :
https://github.com/foldable-devices/foldable-windows-registry-specification

When/If Windows receive a proper native API we will remove this approach
in favor of the official one.

Bug: 1039050
Change-Id: I3d949fbca5842758cd9c197fdf02c7d5838766ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4671552
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172359}
NOKEYCHECK=True
GitOrigin-RevId: fc28a0b39cad37a2a12295d5142c3f013a5384c2