Prerender: Cancel on activation mismatch due to auxiliary browsing contexts
This CL cancels prerendering on activation mismatch due to the existence
of auxiliary browsing contexts with a new PrerenderFinalStatus. This
cancellation should be ok as a trigger page will be navigated away and
prerendered pages will eventually be destroyed.

This final status should be useful for web developers. Before this CL,
this activation mismatch was not shown anywhere, so they didn't know the
reason why prerendered pages were not used.

Bug: 1459901
Change-Id: I5f2d628f3a1d40672a060f377c43d01fd530a700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4699922
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Huanpo Lin <robertlin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172653}
NOKEYCHECK=True
GitOrigin-RevId: 2f42734d2a955b1c8648eba0617f9d694f4786be