[Autofill] Add a new field is_password_type_by_heuristics.
With this field new heuristics for password input recognition which
include, but are not limited to, recognition of key words such as
password and its related translations and abbreviations in the 'id'
or 'name' of the field are introduced. Context of the field is recognized as password without the field necessarily being of type "password" in the HTML code.

Bug: 1466053
TODO: Add and improve tests, improve heuristics.
Change-Id: Ibd09b716623dd94ed2acc730c7a9fb1691e72cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4693902
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172834}
NOKEYCHECK=True
GitOrigin-RevId: 80ca9527e4974011067ac27383c62a82d4755e1a