Error close_promise of outgoing stream on error
If an outgoing stream from webtransport is closed,
a close promise is generated and a STOP_SENDING is send.
During the time of closing the stream, e.g. a session close can error
 the stream, causing a close promise, that is never settled.
This patch addresses this issue and closes the promise with error.
Symptom of the problem is an outgoing stream close hanging forever.

R=ricea@chromium.org

Bug: 1455664
Change-Id: I9ea5c644095b514bec37383084fa4e389ebcbd3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697154
Auto-Submit: Marten Richter <marten.richter@tu-berlin.de>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172946}
NOKEYCHECK=True
GitOrigin-RevId: d6b544ac01b52381a82fc88d3a28d27642c61cd0